fix: grouped checklists bug fixes and basic editor-side test (#4144)
