+++ editor.planx.uk/src/@planx/components/Checklist/Editor/components/ExclusiveOrOptionManager.tsx
@@ -35,16 +35,21 @@ export const ExclusiveOrOptionManager = ({
             values={exclusiveOptions || []}
             onChange={(newExclusiveOptions) => {
               if (grouped) {
-                const exclusiveOptionGroup = {
-                  title: "Or",
-                  exclusive: true,
-                  children: newExclusiveOptions,
+                const updateGroupedOptions = () => {
+                  if (newExclusiveOptions.length > 0) {
+                    const exclusiveOptionGroup = {
+                      title: "Or",
+                      exclusive: true,
+                      children: newExclusiveOptions,
+                    };
+
+                    return [...nonExclusiveOptions, exclusiveOptionGroup];
+                  } else {
+                    return [...nonExclusiveOptions];
+                  }
                 };
-                const newCombinedOptions = [
-                  ...nonExclusiveOptions,
-                  exclusiveOptionGroup,
-                ];
-                formik.setFieldValue("groupedOptions", newCombinedOptions);
+
+                formik.setFieldValue("groupedOptions", updateGroupedOptions());
                 return;
               }
               const newCombinedOptions = [
@@ -55,7 +60,7 @@ export const ExclusiveOrOptionManager = ({
             }}
             newValueLabel='add "or" option'
             maxItems={1}
-            disableDragAndDrop
+            noDragAndDrop
             newValue={() => {
               return {
                 data: {
