+++ editor.planx.uk/src/@planx/components/Checklist/Public/components/ExclusiveChecklistItem.tsx
@@ -1,5 +1,7 @@
+import Box from "@mui/material/Box";
 import Grid from "@mui/material/Grid";
 import Typography from "@mui/material/Typography";
+import { visuallyHidden } from "@mui/utils";
 import { FormikProps } from "formik";
 import React from "react";
 import FormWrapper from "ui/public/FormWrapper";
@@ -18,14 +20,18 @@ export const ExclusiveChecklistItem = ({
 }) => (
   <FormWrapper key={exclusiveOrOption.id}>
     <Grid item xs={12} key={exclusiveOrOption.data.text}>
-      <Typography width={36} display="flex" justifyContent="center">
-        or
-      </Typography>
+      <Box width={36}>
+        <Typography align="center">or</Typography>
+      </Box>
+      <p id="exclusive-option-explanation" style={visuallyHidden}>
+        Selecting this option will deselect all other options.
+      </p>
       <ChecklistItem
         onChange={changeCheckbox(exclusiveOrOption.id)}
         label={exclusiveOrOption.data.text}
         id={exclusiveOrOption.id}
         checked={formik.values.checked.includes(exclusiveOrOption.id)}
+        inputProps={{ "aria-describedby": "exclusive-option-explanation" }}
       />
     </Grid>
   </FormWrapper>
